refactor(server) #4895: compose the OpenAPI spec from per-domain contributors, attach bearerAuth at the root
Security moves from a per-operation basicAuth requirement to a root-level
basicAuth-or-bearerAuth declaration, which is what AbstractServerHttpHandler
actually accepts on every authenticated route. Liveness and readiness override
it with an empty requirement.

Removes OpenApiHandler and OpenApiDocsHandler: neither was ever registered in
HttpServer, and both survived only because two tests asserted their existence
by reflection.
