--- server/src/main/java/com/arcadedb/server/http/handler/OpenApiDocsHandler.java
@@ -1,32 +1,0 @@
-/*
- * Copyright © 2021-present Arcade Data Ltd (info@arcadedata.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-FileCopyrightText: 2021-present Arcade Data Ltd (info@arcadedata.com)
- * SPDX-License-Identifier: Apache-2.0
- */
-package com.arcadedb.server.http.handler;
-
-import com.arcadedb.server.http.HttpServer;
-
-/**
- * Alias class for GetApiDocsHandler to maintain backward compatibility.
- * This class exists solely to provide a consistent naming convention for handler discovery.
- */
-public class OpenApiDocsHandler extends GetApiDocsHandler {
-
-  public OpenApiDocsHandler(final HttpServer httpServer) {
-    super(httpServer);
-  }
-}
