--- server/src/main/java/com/arcadedb/server/http/handler/OpenApiHandler.java
@@ -1,33 +1,0 @@
-/*
- * Copyright © 2021-present Arcade Data Ltd (info@arcadedata.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-FileCopyrightText: 2021-present Arcade Data Ltd (info@arcadedata.com)
- * SPDX-License-Identifier: Apache-2.0
- */
-package com.arcadedb.server.http.handler;
-
-import com.arcadedb.server.http.HttpServer;
-
-/**
- * Alias for GetOpenApiHandler to match test expectations.
- * This class extends GetOpenApiHandler to provide backward compatibility
- * with test expectations while maintaining consistent naming conventions.
- */
-public class OpenApiHandler extends GetOpenApiHandler {
-
-  public OpenApiHandler(final HttpServer httpServer) {
-    super(httpServer);
-  }
-}
